Return rendering tests to main test file
