+++ packages/perseus/src/widgets/orderer/orderer.test.ts
@@ -1,3 +1,5 @@
+import {act} from "@testing-library/react";
+
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
 import {renderQuestion} from "../__testutils__/renderQuestion";
@@ -37,5 +39,50 @@ describe("orderer widget", () => {
 
         // Assert
         expect(container).toMatchSnapshot("first mobile render");
+    });
+
+    it("can be answered correctly", () => {
+        // Arrange
+        const apiOptions: APIOptions = {
+            isMobile: false,
+        };
+        const {renderer} = renderQuestion(question2, apiOptions);
+        const [orderer] = renderer.findWidgets("orderer 1");
+
+        // Act
+        act(() => orderer.setListValues(["1", "2", "3"]));
+
+        // assert
+        expect(renderer).toHaveBeenAnsweredCorrectly();
+    });
+
+    it("can be answered incorrectly", () => {
+        // Arrange
+        const apiOptions: APIOptions = {
+            isMobile: false,
+        };
+        const {renderer} = renderQuestion(question2, apiOptions);
+        const [orderer] = renderer.findWidgets("orderer 1");
+
+        // Act
+        act(() => orderer.setListValues(["3", "2", "1"]));
+
+        // assert
+        expect(renderer).toHaveBeenAnsweredIncorrectly();
+    });
+
+    it("is invalid when no options are selected", () => {
+        // Arrange
+        const apiOptions: APIOptions = {
+            isMobile: false,
+        };
+        const {renderer} = renderQuestion(question2, apiOptions);
+        const [orderer] = renderer.findWidgets("orderer 1");
+
+        // Act
+        act(() => orderer.setListValues([]));
+
+        // assert
+        expect(renderer).toHaveInvalidInput();
     });
 });
